requirements.txt
Django==1.11.2
django-ajax-selects==1.6.0
django-appconf==1.0.2
django-bootstrap3==8.2.3
django-braces==1.11.0
django-crispy-forms==1.6.1
django-datatables-view==1.14.0
django-date-extensions==2.0
django-extra-views==0.9.0
django-fm==0.2.5
django-generic-json-views==0.8
Django-Select2==5.10.0
django-summernote==0.8.7
django-widget-tweaks==1.4.1
MySQL-python==1.2.5
pytz==2017.2
PyYAML==3.11
six==1.10.0
uuid==1.30
-e git://github.com/JoshBone/django_cloneable.git#egg=django_cloneable
-e git://github.com/JoshBone/django-extra-views.git#egg=extra-views
django-userena==2.0.1